Fix database credentials config
Turns out the problem I had with application.conf file not reading env variables
was because I was using a wrong syntax.
${A_VARIABLE} is not supported in Play, or it is used in some other way.
And it means Intellij can insert environment variables into run configurations
just fine.
Also, if you use the "start.of.var"{?END_OF_VAR} syntax, play interprets it as
if the end of var is optional, and "start.of.var" can be used on its own,
which is not the case for postgres jdbc url.